% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine.query.results}
\alias{combine.query.results}
\title{combine.query.results}
\usage{
combine.query.results(query, genes, new.name)
}
\arguments{
\item{query}{The result of perform.query, perform.query.timeseries of perform.query.from.region}

\item{genes}{A list of genes to be merged}

\item{new.name}{A string containing the new name to be used for the new genes}
}
\value{
The frequencies of the genes in the cancer related documents and in all the documents retireved on PubMed
}
\description{
Merge a set of genes in a unique gene
}
\examples{
data(query)
combine.query.results(query, c('IDH1', 'IDH2'), 'new_gene')

}

