% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine.single.matrix}
\alias{combine.single.matrix}
\title{combine.single.matrix}
\usage{
combine.single.matrix(query, genes, new.name)
}
\arguments{
\item{query}{The result of perform.query, perform.query.timeseries of perform.query.from.region}

\item{genes}{A list of genes to be merged}

\item{new.name}{A string containing the new name to be used for the new genes}
}
\value{
a merged matrix
}
\description{
Perform merge procedure on a matrix
}
