% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoScore.R
\name{compute.frequencies.scores}
\alias{compute.frequencies.scores}
\title{compute.frequencies.scores}
\usage{
compute.frequencies.scores(data, filter.threshold = 1,
  analysis.mode = "Log2")
}
\arguments{
\item{data}{input data as result of the function perform.query}

\item{filter.threshold}{threshold to filter for a minimum number of citations for the genes}

\item{analysis.mode}{logaritmic scores to be computed, i.e., log10, log2, natural log or log5}
}
\value{
the computed scores
}
\description{
compute the logaritmic scores based on the frequencies of the genes
}
