% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoScore.R
\name{compute.OncoScore.from.region}
\alias{compute.OncoScore.from.region}
\title{compute.OncoScore.from.region}
\usage{
compute.OncoScore.from.region(chromosome, start = NA, end = NA,
  gene.num.limit = 100, filter.threshold = NA, analysis.mode = "Log2",
  cutoff.threshold = 21.09, file = NULL)
}
\arguments{
\item{chromosome}{TODO}

\item{start}{TODO}

\item{end}{TODO}

\item{gene.num.limit}{TODO}

\item{filter.threshold}{TODO}

\item{analysis.mode}{TODO}

\item{cutoff.threshold}{TODO}

\item{file}{TODO}
}
\value{
the computed scores
}
\description{
TODO
}

