% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoScore.R
\name{compute.oncoscore}
\alias{compute.oncoscore}
\title{compute.oncoscore}
\usage{
compute.oncoscore(data, filter.threshold = 0, analysis.mode = "Log2",
  cutoff.threshold = 21.09, file = NULL, filter.invalid = TRUE)
}
\arguments{
\item{data}{input data as result of the function perform.query}

\item{filter.threshold}{threshold to filter for a minimum number of citations for the genes}

\item{analysis.mode}{logaritmic scores to be computed, i.e., log10, log2, natural log or log5}

\item{cutoff.threshold}{threshold to be used to asses the oncogenes}

\item{file}{should I save the results to text files?}

\item{filter.invalid}{auto-remove genes with invalid count}
}
\value{
the computed OncoScores and the clusters for the genes
}
\description{
compute the OncoScore for a list of genes
}
\examples{
data(query)
compute.oncoscore(query)

}

