% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoScore.R
\name{compute.oncoscore.timeseries}
\alias{compute.oncoscore.timeseries}
\title{compute.oncoscore.timeseries}
\usage{
compute.oncoscore.timeseries(
  data,
  filter.threshold = 0,
  analysis.mode = "Log2",
  cutoff.threshold = 21.09,
  file = NULL
)
}
\arguments{
\item{data}{input data as result of the function perform.query.timeseries}

\item{filter.threshold}{threshold to filter for a minimum number of citations for the genes}

\item{analysis.mode}{logaritmic scores to be computed, i.e., log10, log2, natural log or log5}

\item{cutoff.threshold}{threshold to be used to asses the oncogenes}

\item{file}{should I save the results to text files?}
}
\value{
the performed OncoScores time series analysis
}
\description{
perform the OncoScore time series analysis for a list of genes and data times
}
\examples{
data(query.timepoints)
compute.oncoscore.timeseries(query.timepoints)

}
