% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoScore.R
\name{estimate.oncogenes}
\alias{estimate.oncogenes}
\title{estimate.oncogenes}
\usage{
estimate.oncogenes(data, cutoff.threshold = 21.09)
}
\arguments{
\item{data}{input data as result of the function compute.frequencies.scores}

\item{cutoff.threshold}{threshold to be used to asses the oncogenes}
}
\value{
the computed scores and oncogenes
}
\description{
estimate the oncoscore for the genes
}
