% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{get.driver.file}
\alias{get.driver.file}
\title{get.driver.file}
\usage{
get.driver.file(getURL, showurl = FALSE, sep = "\\n", quiet = TRUE,
  clean = FALSE, verbose = FALSE)
}
\arguments{
\item{getURL}{TODO}

\item{showurl}{TODO}

\item{sep}{TODO}

\item{quiet}{TODO}

\item{clean}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
get the URL webpage
}

