% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.pubmed.driver.analysis}
\alias{get.pubmed.driver.analysis}
\title{get.pubmed.driver.analysis}
\usage{
get.pubmed.driver.analysis(keywords, gene)
}
\arguments{
\item{keywords}{The set of keywords to be used for the query to PubMed}

\item{gene}{The name of a gene to be used for the query to PubMed}
}
\value{
The frequency for the current gene retrieved with the query on the provided set of keywords
}
\description{
query PubMed for a list of genes
}

