% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{getListFromXMLdriver}
\alias{getListFromXMLdriver}
\title{DelayDriver}
\usage{
getListFromXMLdriver(webget, smt = FALSE, sme = FALSE, MaxRet = 3e+06,
  return.data = FALSE)
}
\arguments{
\item{webget}{TODO}

\item{smt}{TODO}

\item{sme}{TODO}

\item{MaxRet}{TODO}

\item{return.data}{TODO}
}
\value{
TODO
}
\description{
process the result of the query
}

