% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{GetPubMedDriverAnalysis}
\alias{GetPubMedDriverAnalysis}
\title{GetPubMedDriverAnalysis}
\usage{
GetPubMedDriverAnalysis(keywords, gene, showurl = FALSE, MaxRet = 3e+06,
  sme = FALSE, smt = FALSE)
}
\arguments{
\item{keywords}{TODO}

\item{gene}{TODO}

\item{showurl}{TODO}

\item{MaxRet}{TODO}

\item{sme}{TODO}

\item{smt}{TODO}
}
\value{
TODO
}
\description{
query PubMed for a list of genes
}

