% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{ncbi2r.driver.options.default}
\alias{ncbi2r.driver.options.default}
\title{ncbi2r.driver.options.default}
\usage{
ncbi2r.driver.options.default()
}
\value{
TODO
}
\description{
setup the default parameters for the queries to PubMed
}

