% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{ncbi2r.driver.options}
\alias{ncbi2r.driver.options}
\title{ncbi2r.driver.options}
\usage{
ncbi2r.driver.options()
}
\value{
TODO
}
\description{
setup the parameters for the queries to PubMed to a global variable
}

