% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoScore.R
\name{perform.OncoScore.TimeSeries}
\alias{perform.OncoScore.TimeSeries}
\title{perform.OncoScore.TimeSeries}
\usage{
perform.OncoScore.TimeSeries(data, filter.threshold = NA,
  analysis.mode = "Log2", cutoff.threshold = 21.09, make.report = TRUE,
  results.file = "Results_TimeSeries_OncoScore.txt")
}
\arguments{
\item{data}{input data as result of the function perform.time.series.query}

\item{filter.threshold}{threshold to filter for a minimum number of citations for the genes}

\item{analysis.mode}{logaritmic scores to be computed, i.e., log10, log2, natural log or log5}

\item{cutoff.threshold}{threshold to be used to asses the oncogenes}

\item{make.report}{should I save the results to text files?}

\item{results.file}{file where to save the results}
}
\value{
the performed OncoScores time series analysis
}
\description{
perform the OncoScore time series analysis for a list of genes and data times
}
\examples{

}

