% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{perform.query.from.region}
\alias{perform.query.from.region}
\title{perform.query.from.region}
\usage{
perform.query.from.region(chromosome, start = NA, end = NA,
  gene.num.limit = 100)
}
\arguments{
\item{chromosome}{chromosome to be retireved}

\item{start}{initial position to be used}

\item{end}{final position to be used}

\item{gene.num.limit}{A limit to the genes to be considered in the analysis; this is done to limit the number of queries to PubMed}
}
\value{
The frequencies of the genes in the cancer related documents and in all the documents retireved on PubMed
}
\description{
Perform the query to PubMed on a given chromosomic region
}
\examples{
chromosome = 15
start = 200000
end = 300000
\donttest{perform.query.from.region(chromosome, start, end)}

}
