% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{perform.query.timeseries}
\alias{perform.query.timeseries}
\title{perform.query.timeseries}
\usage{
perform.query.timeseries(list.of.genes, list.of.datatimes,
  gene.num.limit = 100, timepoints.limit = 10, custom.search = NA)
}
\arguments{
\item{list.of.genes}{The list of genes to be used in the queries to PubMed}

\item{list.of.datatimes}{The list of time points to be used in the queries to PubMed}

\item{gene.num.limit}{A limit to the genes to be considered in the analysis; this is done to limit the number of queries to PubMed}

\item{timepoints.limit}{A limit to the time points to be considered in the analysis; this is done to limit the number of queries to PubMed}

\item{custom.search}{A custom set of keyworkds to be used when quering PubMed}
}
\value{
The frequencies of the genes in the cancer related documents and in all the documents retireved on PubMed at the specified time points
}
\description{
perforn the query to PubMed for the time series analysis
}
\examples{
data(genes)
data(timepoints)
\donttest{perform.query.timeseries(genes[1:2], timepoints[1:2])}

}

