% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.PubMed.R
\name{perform.time.series.query}
\alias{perform.time.series.query}
\title{perform.time.series.query}
\usage{
perform.time.series.query(list.of.genes, list.of.datatimes,
  gene.num.limit = 100, timepoints.limit = 10)
}
\arguments{
\item{list.of.genes}{TODO}

\item{list.of.datatimes}{TODO}

\item{gene.num.limit}{TODO}

\item{timepoints.limit}{TODO}
}
\value{
TODO
}
\description{
perforn the query to PubMed for the time series analysis
}
\examples{

}

