% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{perform.web.query}
\alias{perform.web.query}
\title{perform.web.query}
\usage{
perform.web.query(list.of.genes, gene.num.limit = 100, custom.search = NA)
}
\arguments{
\item{list.of.genes}{The list of genes to be used in the queries to PubMed}

\item{gene.num.limit}{A limit to the genes to be considered in the analysis; this is done to limit the number of queries to PubMed}

\item{custom.search}{A custom set of keyworkds to be used when quering PubMed}
}
\value{
The frequencies of the genes in the cancer related documents and in all the documents retireved on PubMed
}
\description{
perforn the query to PubMed
}
\examples{
data(genes)
\donttest{perform.web.query(genes[1:2])}

}

