% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oncoscore}
\alias{plot.oncoscore}
\title{plot.oncoscore}
\usage{
\method{plot}{oncoscore}(data, gene.number = 5, ...)
}
\arguments{
\item{data}{input data as result of the function compute.OncoScore}

\item{gene.number}{number of genes to print}

\item{...}{additional parameter to pass to the barplot function}
}
\description{
plot the OncoScore for a list of genes
}
\examples{
data(query)
result = compute.OncoScore(query)
plot.oncoscore(result)

}

