% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oncoscore.timeseries}
\alias{plot.oncoscore.timeseries}
\title{plot.oncoscore.timeseries}
\usage{
\method{plot}{oncoscore.timeseries}(x, gene.number = 5, main = "OncoScore",
  xlab = "score", ylab = "genes", ...)
}
\arguments{
\item{x}{input data as result of the function compute.OncoScore}

\item{gene.number}{number of genes to print}

\item{main}{the title}

\item{xlab}{description of x asix (defaul score)}

\item{ylab}{description of y asix (defaul genes)}

\item{...}{additional parameter to pass to the lines function}
}
\description{
plot the OncoScore for a list of genes
}
\examples{
data(query.timepoints)
result = compute.OncoScore.TimeSeries(query.timepoints)
plot.oncoscore.timeseries(result)

}

