% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query}
\alias{query}
\title{The result of perform.web.query on genes}
\format{rdata}
\source{
fake data
}
\usage{
data(query)
}
\description{
This dataset contains the result of perform.web.query on genes
}
\author{
Luca De Sano
}

