% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{showMessages}
\alias{showMessages}
\title{showMessages}
\usage{
showMessages(webget, smt = FALSE, sme = FALSE)
}
\arguments{
\item{webget}{TODO}

\item{smt}{TODO}

\item{sme}{TODO}
}
\value{
TODO
}
\description{
show error and warnings messages to standard output
}

