% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{splitfirst}
\alias{splitfirst}
\title{splitfirst}
\usage{
splitfirst(lineoftext, searchfor = " ", characterposition = 1)
}
\arguments{
\item{lineoftext}{TODO}

\item{searchfor}{TODO}

\item{characterposition}{TODO}
}
\value{
TODO
}
\description{
utility function to split items
}

