% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timepoints}
\alias{timepoints}
\title{A list of timepoints}
\format{txt}
\source{
example data
}
\usage{
data(timepoints)
}
\description{
This dataset contains a list of time points to be used in the analysis as an example
}
\author{
Luca De Sano
}

