% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{tryScan}
\alias{tryScan}
\title{tryScan}
\usage{
tryScan(getURL, sep = "\\n", quiet = TRUE, retry = 20, error = TRUE,
  verbose = FALSE)
}
\arguments{
\item{getURL}{TODO}

\item{sep}{TODO}

\item{quiet}{TODO}

\item{retry}{TODO}

\item{error}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
try to query the given URL
}

