% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoScore.R
\name{compute.OncoScore}
\alias{compute.OncoScore}
\title{compute.OncoScore}
\usage{
compute.OncoScore(data, filter.threshold = NA, analysis.mode = "Log2",
  cutoff.threshold = 21.09, file = NULL)
}
\arguments{
\item{data}{input data as result of the function perform.web.query}

\item{filter.threshold}{threshold to filter for a minimum number of citations for the genes}

\item{analysis.mode}{logaritmic scores to be computed, i.e., log10, log2, natural log or log5}

\item{cutoff.threshold}{threshold to be used to asses the oncogenes}

\item{make.report}{should I save the results to text files?}

\item{results.file}{file where to save the results}
}
\value{
the computed OncoScores and the clusters for the genes
}
\description{
compute the OncoScore for a list of genes
}
\examples{

}

