% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{A list of genes}
\format{txt}
\source{
example data
}
\usage{
data(genes)
}
\description{
This dataset contains a list of genes to be used in the analysis as an example
}
\author{
Luca De Sano
}

