% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.genes.from.biomart}
\alias{get.genes.from.biomart}
\title{get.genes.from.biomart}
\usage{
get.genes.from.biomart(chromosome, start = NA, end = NA)
}
\arguments{
\item{chromosome}{chromosome to be retireved}

\item{start}{initial position to be used}

\item{end}{final position to be used}
}
\description{
Get a gene list from biomart
}
\examples{
\donttest{ch15 = get.genes.from.biomart(15, 200000, 300000)}

}

