% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.list.from.xml}
\alias{get.list.from.xml}
\title{get.list.from.xml}
\usage{
get.list.from.xml(webget)
}
\arguments{
\item{webget}{The result from the query to PubMed}
}
\value{
Processed result obtained from the query to PubMed
}
\description{
process the result of the query
}

