% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.PubMed.R
\name{get.pubmed.driver.analysis}
\alias{get.pubmed.driver.analysis}
\title{get.pubmed.driver.analysis}
\usage{
get.pubmed.driver.analysis(keywords, gene)
}
\arguments{
\item{keywords}{TODO}

\item{gene}{TODO}
}
\value{
TODO
}
\description{
query PubMed for a list of genes
}

