% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.PubMed.R
\name{perform.web.query}
\alias{perform.web.query}
\title{perform.web.query}
\usage{
perform.web.query(list.of.genes, gene.num.limit = 100, custom.search = NA)
}
\arguments{
\item{list.of.genes}{TODO}

\item{gene.num.limit}{TODO}

\item{custom.search}{TODO}
}
\value{
TODO
}
\description{
perforn the query to PubMed
}
\examples{
data(genes)
perform.web.query(genes[1:2])

}

