% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oncoscore}
\alias{plot.oncoscore}
\title{plot.oncoscore}
\usage{
\method{plot}{oncoscore}(x, gene.number = 5, main = "OncoScore",
  xlab = "score", ylab = "genes", file = NA, ...)
}
\arguments{
\item{x}{input data as result of the function compute.OncoScore}

\item{gene.number}{number of genes to print}

\item{main}{the title}

\item{xlab}{description of x asix (defaul score)}

\item{ylab}{description of y asix (defaul genes)}

\item{file}{where to save the plot}

\item{...}{additional parameter to pass to the barplot function}
}
\value{
A plot
}
\description{
plot the OncoScore for a list of genes
}
\examples{
data(query)
result = compute.oncoscore(query)
plot.oncoscore(result)

}
