% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oncoscore.timeseries}
\alias{plot.oncoscore.timeseries}
\title{plot.oncoscore.timeseries}
\usage{
\method{plot}{oncoscore.timeseries}(
  x,
  gene.number = 5,
  incremental = FALSE,
  relative = FALSE,
  main = "OncoScore",
  xlab = "timepoints",
  ylab = "score",
  legend.pos = "top",
  file = NA,
  ...
)
}
\arguments{
\item{x}{input data as result of the function compute.OncoScore}

\item{gene.number}{number of genes to print}

\item{incremental}{display the OncoScore increment}

\item{relative}{dispaly the incrementa as relative value}

\item{main}{the title}

\item{xlab}{description of x asix (defaul score)}

\item{ylab}{description of y asix (defaul genes)}

\item{legend.pos}{Position of the legend}

\item{file}{where to save the plot}

\item{...}{additional parameter to pass to the lines function}
}
\value{
A plot
}
\description{
plot the OncoScore for a list of genes
}
\examples{
data(query.timepoints)
result = compute.oncoscore.timeseries(query.timepoints)
plot.oncoscore.timeseries(result)

}
