% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query.timepoints}
\alias{query.timepoints}
\title{The result of perform.time.series.query on genes and timepoints}
\format{
rdata
}
\source{
example data
}
\usage{
query.timepoints
}
\value{
list of 5 matrix 5 x 2
}
\description{
This dataset contains the result of perform.time.series.query on genes and timepoints
}
\keyword{datasets}
