% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query.timepoints}
\alias{query.timepoints}
\title{The result of perform.time.series.query on genes and timepoints}
\format{rdata}
\source{
fake data
}
\usage{
data(query.timepoints)
}
\description{
This dataset contains the result of perform.time.series.query on genes and timepoints
}
\author{
Luca De Sano
}

